+++ packages/cli/src/util/response-error.ts
@@ -1,4 +1,4 @@
-import type { Response } from 'node-fetch';
+import type { Response } from './fetch';
 import { APIError } from './errors-ts';
 
 export default async function responseError(