+++ packages/cli/test/unit/commands/link/index.test.ts
@@ -1037,15 +1037,14 @@ describe('link', () => {
     expect(exitCode, 'exit code for "link"').toEqual(0);
 
     expect(await readJSON(join(cwd, 'vercel.json'))).toMatchObject({
-      experimentalServices: {
+      services: {
         frontend: {
+          root: '.',
           framework: 'nextjs',
-          routePrefix: '/',
         },
         api: {
           root: 'services/api',
           entrypoint: 'index:app',
-          routePrefix: '/_/api',
         },
       },
     });
@@ -1251,15 +1250,14 @@ describe('link', () => {
     expect(exitCode, 'exit code for "link"').toEqual(0);
 
     expect(await readJSON(join(cwd, 'apps/web/vercel.json'))).toMatchObject({
-      experimentalServices: {
+      services: {
         frontend: {
+          root: '.',
           framework: 'nextjs',
-          routePrefix: '/',
         },
         api: {
           root: 'services/api',
           entrypoint: 'index:app',
-          routePrefix: '/_/api',
         },
       },
     });
@@ -1330,15 +1328,14 @@ describe('link', () => {
     expect(exitCode, 'exit code for "link"').toEqual(0);
 
     expect(await readJSON(join(cwd, 'apps/web/vercel.json'))).toMatchObject({
-      experimentalServices: {
+      services: {
         frontend: {
+          root: '.',
           framework: 'nextjs',
-          routePrefix: '/',
         },
         api: {
           root: 'services/api',
           entrypoint: 'index:app',
-          routePrefix: '/_/api',
         },
       },
     });