+++ packages/fs-detectors/test/fixtures/e2e/09-services-frontend-backend-go-zc/frontend/app/not-found.js
@@ -1,3 +1,0 @@
-export default function NotFound() {
-  return '404 from Next.js';
-}